Put isEdit isSearched into seperate file
