+++ source/scripts/ManageRecipe.js
@@ -1,4 +1,5 @@
 import {storage} from './storage.js';
+import {isEdit, isSearched} from './url.js';
 
 window.addEventListener('DOMContentLoaded', init);
 
@@ -299,28 +300,3 @@ firstInstr.addEventListener('keydown', (event) => {
     firstInstr.nextSibling.focus();
   }
 });
-
-/**
- * Check if we are editing a recipe
- * @return {String} the editing recipe id if editing, else '-1'
- */
-function isEdit() {
-  // console.log('storage:' + storage);
-  // Extract query id
-  const queryString = window.location.search;
-  const urlParams = new URLSearchParams(queryString);
-  const id = urlParams.get('id');
-  return id != null ? id : '-1';
-}
-
-/**
- * Check if we are editing from a searched recipe
- * @return {Boolean} true if yes, else false
- */
-function isSearched() {
-  // console.log('storage:' + storage);
-  // Extract query id
-  const queryString = window.location.search;
-  const urlParams = new URLSearchParams(queryString);
-  return urlParams.get('searched') != null;
-}
